requirements.txt
fastapi==0.82.0
pydantic==1.9.2
python-jose==3.3.0
uvicorn==0.18.3
passlib==1.7.4
bcrypt==3.2.0
python-dotenv==0.21.0